@@ -10,10 +10,12 @@ const main = async () => {
   const consumer = new KafkaALOConsumer({
     'group.id': 'alo-consumer-test-1',
     'metadata.broker.list': '127.0.0.1:9092',
-    'auto.offset.reset': 'largest',
     'enable.auto.offset.store': false,
     'enable.auto.commit': false,
-  }, {})
+  }, {
+      'auto.offset.reset': 'largest',
+    })
+
   await consumer.connect()
   await consumer.subscribe([
     'rdkafka-test0',
