@@ -1,22 +1,26 @@
-import * as _ from 'lodash'
+/* tslint:disable */
 import * as Kafka from 'node-rdkafka'
+import * as _ from 'lodash'
 import * as bluebird from 'bluebird'
 
 import { TopicPartition, KafkaMetadata, KafkaMessage, KafkaMessageError } from './types'
 import {
   ConnectingError,
   DisconnectError,
-  ConnectionNotReadyError,
   ConnectionDeadError,
   ConsumerRuntimeError,
   MetadataError,
   SeekError,
 } from './errors'
 
-const SEEK_TIMEOUT = 1000
+const DEFAULT_CONSUME_SIZE = 100
+const DEFAULT_CONCURRENT = 100
+const DEFAULT_SEEK_TIMEOUT = 1000
 const ErrorCode = Kafka.CODES.ERRORS
 
-const _ifNotExistedAndSet = (conf: any, key: string, value: any) => {
+type Store = { [key: string]: { [key: string]: number } }
+
+const _setIfNotExist = (conf: any, key: string, value: any) => {
   if (conf[key] === undefined) {
     conf[key] = value
     return true
@@ -26,30 +30,27 @@ const _ifNotExistedAndSet = (conf: any, key: string, value: any) => {
 
 export abstract class KafkaBasicConsumer {
   public consumer: Kafka.KafkaConsumer
-  protected dead: boolean
+  protected dying: boolean
   protected topics: string[]
-
-  protected offsetStore: { [key: string]: { [key: number]: number } } = {}
-  protected errOffsetStore: { [key: string]: { [key: number]: number } } = {}
+  // store the largest offset that has been success consumed in each topic-partition
+  protected offsetStore: Store = {}
+  protected errOffsetStore: Store = {}
 
   constructor(conf: any, topicConf: any = {}) {
-    this.dead = false
+    this.dying = false
     this.topics = []
 
-    _ifNotExistedAndSet(conf, 'rebalance_cb', (err: any, assignment: any) => {
+    _setIfNotExist(conf, 'rebalance_cb', (err: any, assignment: any) => {
       if (err.code === ErrorCode.ERR__ASSIGN_PARTITIONS) {
-        // Note: this can throw when you are disconnected. Take care and wrap it in
-        // a try catch if that matters to you
         this.consumer.assign(assignment)
-        console.log(`Consumer rebalanced at : `)
+        let rebalance_log = 'consumer rebalance : '
         for (const assign of assignment) {
-          console.log(`   topic ${assign.topic}, partition: ${assign.partition}`)
+          rebalance_log += `{topic ${assign.topic}, partition: ${assign.partition}} `
         }
+        console.log(rebalance_log)
       } else if (err.code == ErrorCode.ERR__REVOKE_PARTITIONS) {
-        // Same as above
         this.consumer.unassign()
       } else {
-        // We had a real error
         console.error(err)
       }
     })
@@ -63,20 +64,19 @@ export abstract class KafkaBasicConsumer {
 
   disconnect() {
     return new Promise((resolve, reject) => {
-      return this.consumer.disconnect((err, data) => {
+      return this.consumer.disconnect((err: Error, data: any) => {
         if (err) {
           reject(new DisconnectError(err.message))
         }
-        console.log('Consumer disconnect success')
+        console.log('consumer disconnect')
         resolve(data)
       })
     })
   }
 
-  // rebalancing is managed internally by librdkafka by default
   async connect(metadataOptions: any = {}) {
     return new Promise((resolve, reject) => {
-      this.consumer.connect(metadataOptions, (err, data) => {
+      this.consumer.connect(metadataOptions, (err: Error, data: any) => {
         if (err) {
           reject(new ConnectingError(err.message))
         }
@@ -88,13 +88,11 @@ export abstract class KafkaBasicConsumer {
 
   private setGracefulDeath() {
     const _gracefulDeath = async () => {
-      console.log('Consumer graceful death begin')
-
-      this.dead = true
+      this.dying = true
       await this.gracefulDead()
       await this.disconnect()
 
-      console.log('Consumer graceful death success')
+      console.log('consumer graceful died')
       process.exit(0)
     }
     process.on('SIGINT', _gracefulDeath)
@@ -106,24 +104,11 @@ export abstract class KafkaBasicConsumer {
     this.topics = _.uniq(_.concat(topics, this.topics))
     // synchronously
     this.consumer.subscribe(this.topics)
-    // refresh offset
-    await this.initOffsetStore()
   }
 
   unsubscribe() {
     this.topics.length = 0
     this.consumer.unsubscribe()
-  }
-
-  getMetadata(metadataOptions: any): Promise<KafkaMetadata> {
-    return new Promise((resolve, reject) => {
-      this.consumer.getMetadata(metadataOptions, (err: Error, data: KafkaMetadata) => {
-        if (err) {
-          reject(new MetadataError(err.message))
-        }
-        resolve(data)
-      })
-    })
   }
 
   seek(toppar: TopicPartition, timeout: number) {
@@ -136,52 +121,6 @@ export abstract class KafkaBasicConsumer {
       })
     })
   }
-
-  async initOffsetStore() {
-    const meta = await this.getMetadata({ timeout: 1000 })
-    for (const topic of meta.topics) {
-      if (this.topics.includes(topic.name)) {
-        this.offsetStore[topic.name] = {}
-        this.errOffsetStore[topic.name] = {}
-        for (const p of topic.partitions) {
-          this.offsetStore[topic.name][p.id] = -1
-          this.errOffsetStore[topic.name][p.id] = -1
-        }
-      }
-    }
-  }
-
-  async commits() {
-    for (const topic in this.offsetStore) {
-      for (const partition in this.offsetStore[topic]) {
-        let offset = this.offsetStore[topic][partition]
-        const errOffset = this.errOffsetStore[topic][partition]
-        let isNeedSeekBack = false
-
-        if (errOffset >= 0) {
-          offset = errOffset - 1
-          // clear errorOffset
-          this.errOffsetStore[topic][partition] = -1
-          isNeedSeekBack = true
-        }
-
-        if (offset < 0) {
-          continue
-        }
-
-        const toppar = {
-          topic,
-          partition: parseInt(partition),
-          offset: offset + 1,
-        }
-        this.consumer.commitSync(toppar)
-        if (isNeedSeekBack) {
-          await this.seek(toppar, SEEK_TIMEOUT)
-        }
-        this.offsetStore[topic][partition] = -1
-      }
-    }
-  }
 }
 
 // `at least once` Consumer
@@ -189,10 +128,43 @@ export abstract class KafkaBasicConsumer {
 // Otherwise, it will to been blocked on the offset where throw Error
 export class KafkaALOConsumer extends KafkaBasicConsumer {
   constructor(conf: any, topicConf: any = {}) {
-    _ifNotExistedAndSet(conf, 'enable.auto.commit', false)
-    _ifNotExistedAndSet(conf, 'enable.auto.offset.store', false)
+    _setIfNotExist(conf, 'enable.auto.commit', false)
+    _setIfNotExist(conf, 'enable.auto.offset.store', false)
 
     super(conf, topicConf)
+  }
+
+  private setErrOffset(topic: string, partition: number, offset: number) {
+    if (!this.errOffsetStore[topic]) {
+      this.errOffsetStore[topic] = {}
+    }
+    if (this.errOffsetStore[topic][partition] === undefined) {
+      this.errOffsetStore[topic][partition] = offset
+    } else {
+      this.errOffsetStore[topic][partition] = Math.min(
+        this.errOffsetStore[topic][partition],
+        offset,
+      )
+    }
+  }
+
+  private setOffset(topic: string, partition: number, offset: number) {
+    if (!this.offsetStore[topic]) {
+      this.offsetStore[topic] = {}
+    }
+
+    this.offsetStore[topic][partition] = Math.max(
+      this.offsetStore[topic][partition] || 0,
+      offset,
+    )
+  }
+
+  private getOffset(topic: string, partition: number | string) {
+    return this.offsetStore[topic] ? this.offsetStore[topic][partition] : undefined
+  }
+
+  private getErrOffset(topic: string, partition: number | string) {
+    return this.errOffsetStore[topic] ? this.errOffsetStore[topic][partition] : undefined
   }
 
   async gracefulDead(): Promise<boolean> {
@@ -200,66 +172,77 @@ export class KafkaALOConsumer extends KafkaBasicConsumer {
     return true
   }
 
+  async getMetadata(metadataOptions: any) {
+    return new Promise<KafkaMetadata>((resolve, reject) => {
+      this.consumer.getMetadata(metadataOptions, (err: Error, data: KafkaMetadata) => {
+        if (err) {
+          reject(new MetadataError(err.message))
+        }
+        resolve(data)
+      })
+    })
+  }
+
+  async commits() {
+    for (const topic of _.keys(this.offsetStore)) {
+      for (const partition of _.keys(this.offsetStore[topic])) {
+        if (this.offsetStore[topic][partition] < 0) {
+          continue
+        }
+        const toppar = {
+          topic,
+          partition: parseInt(partition),
+          offset: this.offsetStore[topic][partition] + 1,
+        }
+        const errOffset = this.getErrOffset(topic, partition)
+        if (errOffset !== undefined) {
+          toppar.offset = errOffset
+          // fallback
+          await this.seek(toppar, DEFAULT_SEEK_TIMEOUT)
+        }
+        this.consumer.commitSync(toppar)
+        console.log(`committed ${JSON.stringify(toppar)}`)
+        this.offsetStore[toppar.topic][toppar.partition] = -1
+        delete this.errOffsetStore[toppar.topic][toppar.partition]
+      }
+    }
+  }
+
   async consume(
     cb: (message: KafkaMessage) => any,
-    options: { size: number, concurrency: number, } = { size: 100, concurrency: 100 },
-  ): Promise<boolean> {
+    options: { size?: number, concurrency?: number, } = {},
+  ): Promise<boolean | KafkaMessage[]> {
     // default option value
-    if (!options.size) {
-      options.size = 100
-    }
-    if (!options.concurrency) {
-      options.concurrency = options.size
-    }
-    let success = true
+    _setIfNotExist(options, 'size', DEFAULT_CONSUME_SIZE)
+    _setIfNotExist(options, 'concurrency', options.size)
 
-    return new Promise<boolean>((resolve, reject) => {
+    return new Promise<boolean | KafkaMessage[]>((resolve, reject) => {
       // This will keep going until it gets ERR__PARTITION_EOF or ERR__TIMED_OUT
       return this.consumer.consume(options.size, async (err: Error, messages: KafkaMessage[]) => {
-        if (this.dead) {
+        if (this.dying) {
           reject(new ConnectionDeadError('Connection has been dead or is dying'))
         }
         if (err) {
           reject(new ConsumerRuntimeError(err.message))
         }
+
         try {
           await bluebird.map(messages, async message => {
-            // stop the topicPartition progress then has error throw
-            if (this.errOffsetStore[message.topic][message.partition] >= 0) {
-              return
-            }
             try {
               await Promise.resolve(cb(message))
-              // stop the topicPartition progress then has error throw
-              if (this.errOffsetStore[message.topic][message.partition] >= 0) {
-                return
-              }
-              // update success offset to max one
-              this.offsetStore[message.topic][message.partition] = Math.max(
-                this.offsetStore[message.topic][message.partition],
-                message.offset,
-              )
-            } catch (e) {
-              success = false
-              // fallback to last message
-              if (this.errOffsetStore[message.topic][message.partition] < 0) {
-                this.errOffsetStore[message.topic][message.partition] = message.offset
-              } else {
-                // fallback to the smallest offset
-                this.errOffsetStore[message.topic][message.partition] = Math.min(
-                  this.errOffsetStore[message.topic][message.partition],
-                  message.offset,
-                )
-              }
+              this.setOffset(message.topic, message.partition, message.offset)
+            } catch (err) {
+              this.setErrOffset(message.topic, message.partition, message.offset)
+              throw err
             }
-          }, { concurrency: options.concurrency })
-
-          await this.commits()
+          }, { concurrency: options.concurrency || DEFAULT_CONCURRENT })
         } catch (e) {
+          await this.commits()
           reject(new ConsumerRuntimeError(e.message))
         }
 
-        return resolve(success)
+        await this.commits()
+        resolve(messages)
       })
     })
   }
@@ -268,9 +251,8 @@ export class KafkaALOConsumer extends KafkaBasicConsumer {
 // `At Most Once` Consumer
 export class KafkaAMOConsumer extends KafkaBasicConsumer {
   constructor(conf: any, topicConf: any = {}) {
-    _ifNotExistedAndSet(conf, 'enable.auto.commit', true)
-    _ifNotExistedAndSet(conf, 'enable.auto.offset.store', true)
-    _ifNotExistedAndSet(conf, 'auto.commit.interval.ms', 500)
+    _setIfNotExist(conf, 'enable.auto.commit', true)
+    _setIfNotExist(conf, 'enable.auto.offset.store', true)
 
     super(conf, topicConf)
   }
@@ -287,26 +269,24 @@ export class KafkaAMOConsumer extends KafkaBasicConsumer {
 
   async consume(
     cb: (message: KafkaMessage) => any,
-    options: { size: number, concurrency: number, } = { size: 100, concurrency: 100 },
-  ): Promise<boolean> {
+    options: { size?: number, concurrency?: number, } = {},
+  ): Promise<boolean | KafkaMessage[]> {
     // default option value
-    if (!options.size) {
-      options.size = 100
-    }
-    if (!options.concurrency) {
-      options.concurrency = options.size
-    }
+    _setIfNotExist(options, 'size', DEFAULT_CONSUME_SIZE)
+    _setIfNotExist(options, 'concurrency', options.size)
+
     let success = true
 
-    return new Promise<boolean>((resolve, reject) => {
+    return new Promise<boolean | KafkaMessage[]>((resolve, reject) => {
       // This will keep going until it gets ERR__PARTITION_EOF or ERR__TIMED_OUT
       return this.consumer.consume(options.size, async (err: Error, messages: KafkaMessage[]) => {
-        if (this.dead) {
+        if (this.dying) {
           reject(new ConnectionDeadError('Connection has been dead or is dying'))
         }
         if (err) {
           reject(new ConsumerRuntimeError(err.message))
         }
+
         try {
           await bluebird.map(messages, async message => {
             try {
@@ -314,7 +294,7 @@ export class KafkaAMOConsumer extends KafkaBasicConsumer {
             } catch (e) {
               success = false
             }
-          }, { concurrency: options.concurrency })
+          }, { concurrency: options.concurrency || DEFAULT_CONCURRENT })
         } catch (e) {
           reject(new ConsumerRuntimeError(err.message))
         }
